requirements.txt
pandas 
pdfplumber 
python-dotenv
google.genai